build: Manually link `ispc_texcomp_astc` to prevent bogus `cargo:rerun-if-changed` lines causing full recompilation
This was added in dcf62dc ("Prepare for release") yet is not an ISPC
module (there's no `ispc_module!(ispc_texcomp_astc)`) nor were the Rust
bindgen bindings added for this library.  It results in `build.rs`
emitting the following line:

    cargo:rerun-if-changed=src/ispc/ispc_texcomp_astc.rs

This file does not exist, causing `cargo` to continuously rebuild this
crate and everything that depends on it (~10 crates on our side).  We
could have removed this library and all logic around it since the only
symbol needed from it resides in a codepath that's commented out, but
MSVC - unlike any other compiler - is unable to deadstrip it and tries
to look for the symbol regardless, resulting in obvious [linker errors].

[linker errors]: https://github.com/Traverse-Research/intel-tex-rs-2/runs/5829771452?check_suite_focus=true