Add `ispc_texcomp_astc` compiled against static Windows/MSVC CRT (#35)
When compiling upstream's `ispc_texcomp_astc.cpp`, `cc-rs` defaults
to using the dynamic MSVC CRT, and this is what the pregenerated
and checked-in `.lib` file uses/requires as well.  This conflicts
with our internal project that sets `target-feature=+crt-static`
(which subsequently only applies to everything that's compiled on
the spot), and appears to be an error of the form `rust-lld: error:
/failifmismatch: mismatch detected for 'RuntimeLibrary'` since Rust
1.86.

To solve this, append the CRT version to the binary name based on
`+crt-static` in `CARGO_CFG_TARGET_FEATURE` when compiling and when
linking the prebuilt, and check the new libraries into the repository.

New binaries are from https://github.com/Traverse-Research/intel-tex-rs-2/actions/runs/16024542690, for all platforms and architectures.